:truck: rename: 각 도메인 폴더들을 domain 패키지 하위로 묶기
